tiaocang ver1

+++ aaas-api-finance/src/main/java/com/shellshellfish/aaas/finance/service/UserProdChangeLogService.java
@@ -13,7 +13,7 @@ public interface UserProdChangeLogService {
 
   List<UserProdChg> getGeneralChangeLogs(Long prodId);
 
-  List<UserProdChgDetail> getDetailChangeLogs(Long prodId, Long changeSeq);
+  List<UserProdChgDetail> getDetailChangeLogs(Long prodId, Integer changeSeq);
 
   boolean insertGeneralChangeLogs(List<UserProdChg> userProdChgs);
